💥  moar testing for release note pick and ordering [b]
BREAKING CHANGE: 💥️ Remove templates and `release-notes` in preference for `release-notes-generator`
