+++ packages/conventional-gitmoji/scripts/copy.sh
@@ -21,5 +21,4 @@ fi
 ###
 # @custom(build)
 ###
-cp -R src/changelog/templates/. ./dist/changelog/templates/
-cp -R src/changelog/templates/. ./dist/templates/
+
