--- packages/conventional-gitmoji/src/changelog/templates/footer.hbs
@@ -1,11 +1,0 @@
-{{#if noteGroups}}
-  {{#each noteGroups}}
-
-#### {{title}}
-
-    {{#each notes}}
-      - {{#if commit.scope}}**{{commit.scope}}**: {{/if}}{{text}}
-    {{/each}}
-  {{/each}}
-
-{{/if}}
