--- packages/conventional-gitmoji/src/changelog/templates/template.hbs
@@ -1,15 +1,0 @@
-{{> header}}
-
-{{#each commitGroups}}
-
-{{#if title}}
-#### {{title}}
-
-{{/if}}
-{{#each commits}}
-{{> commit root=@root}}
-{{/each}}
-
-{{/each}}
-{{> footer}}
-
