+++ packages/conventional-gitmoji/tsup.config.ts
@@ -3,7 +3,7 @@ import { defineConfig, Options } from 'tsup'
 
 import { config as _config } from '../../tsup.config'
 
-const entry = ['src/**', '!src/**/*.hbs']
+const entry = ['src/**']
 const config: Options = {
   ..._config,
   entry,
