--- packages/release-notes/package.json
@@ -1,57 +1,0 @@
-{
-  "name": "@jeromefitz/release-notes",
-  "version": "0.0.0",
-  "description": "Custom Release Notes for Semantic Release.",
-  "author": {
-    "name": "Jerome Fitzgerald",
-    "email": "j@jeromefitzgerald.com",
-    "url": "https://jeromefitzgerald.com"
-  },
-  "license": "MIT",
-  "repository": {
-    "type": "git",
-    "url": "ssh://git@github.com/jeromefitz/packages.git",
-    "directory": "packages/release-notes"
-  },
-  "engines": {
-    "node": ">=14"
-  },
-  "publishConfig": {
-    "registry": "https://registry.npmjs.org/"
-  },
-  "private": false,
-  "scripts": {
-    "build": "tsup",
-    "dev": "tsup --watch",
-    "---": "",
-    "clean": "rm -rf .turbo && rm -rf dist",
-    "clean:install": "yarn clean && rm -rf node_modules",
-    "copy": "mkdir -p ./dist && ./scripts/copy.sh",
-    "--- ": "",
-    "lint:eslint": "eslint ./src --ext cjs,js,jsx,mjs,ts,tsx --max-warnings=0",
-    "lint:prettier": "prettier \"./src/**/*.{cjs,js,jsx,mjs,ts,tsx,json,md,mdx,css,html,yml,yaml,scss}\" --ignore-unknown --loglevel warn",
-    "lint": "yarn lint:prettier --check && yarn lint:eslint",
-    "lint:fix": "yarn lint:eslint --fix && yarn lint:prettier --write",
-    "---  ": "",
-    "semantic-release": "semantic-release",
-    "semantic-release:mono": "semantic-release -e semantic-release-monorepo"
-  },
-  "devDependencies": {
-    "@types/lodash": "4.14.178",
-    "@types/read-pkg-up": "6.0.0"
-  },
-  "dependencies": {
-    "@octokit/core": "3.5.1",
-    "conventional-commits-filter": "2.0.7",
-    "conventional-commits-parser": "3.2.4",
-    "date-fns": "2.28.0",
-    "lodash": "4.17.21",
-    "read-pkg-up": "7.0.1",
-    "semver": "7.3.5"
-  },
-  "type": "module",
-  "sideEffects": false,
-  "typings": "./dist/index.d.ts",
-  "main": "./dist/index.cjs",
-  "module": "./dist/index.js"
-}
