--- packages/release-notes/src/config/index.ts
@@ -1,2 +1,0 @@
-export { HOSTS_CONFIG } from './hosts'
-export { loadChangelogConfig } from './changelog'
