--- packages/release-notes/src/templates/index.ts
@@ -1,4 +1,0 @@
-export { commit } from './commit'
-export { contributor } from './contributor'
-export { footer } from './footer'
-export { header } from './header'
