--- packages/release-notes/tsconfig.json
@@ -1,16 +1,0 @@
-{
-  "compilerOptions": {
-    "esModuleInterop": true,
-    "forceConsistentCasingInFileNames": true,
-    "lib": ["es2020"],
-    "module": "commonjs",
-    "skipLibCheck": true,
-    "strict": true,
-    "target": "es2020",
-    "outDir": "./dist",
-    "rootDir": "./src"
-  },
-  "extends": "@jeromefitz/codestyle/tsconfig.node-14.json",
-  "include": ["src"],
-  "exclude": ["node_modules"]
-}
