--- packages/release-notes/tsup.config.ts
@@ -1,15 +1,0 @@
-// import isCI from 'is-ci'
-import { defineConfig, Options } from 'tsup'
-
-import { config as _config } from '../../tsup.config'
-
-const entry = ['src/*']
-const config: Options = {
-  ..._config,
-  entry,
-  // format: ['esm'],
-}
-
-export default defineConfig({
-  ...config,
-})
