+++ packages/semantic/scripts/copy.sh
@@ -21,4 +21,3 @@ fi
 ###
 # @custom(build)
 ###
-cp -R src/templates/. ./dist/templates/
