--- packages/semantic/src/templates/author.hbs
@@ -1,1 +1,0 @@
-{{#if authorName}}by: **{{authorName}}**{{#if authorEmail}}<{{authorEmail}}>{{/if}}{{/if}}
